Move authenticated to session
Previously whether a client had authenitcated was being recorded on the
AuthController object. This corrects that and now has it residing in the
session. It also moves the checks for whether a session has been
authenticated to the AuthSMTP object so that they are emitted to the
client in a more logical order.

- Adds the AuthSession object to include athenticated on the Session
- Adds check that autherntication happens be DATA, MAIL, RCPT are
  processed by the server.
- Corrects tests so that EHLO or HELO are sent after STARTTLS and
  beofre other commands as needed.

Closes #13