Remove wrap_socket from _trigger_server
Becuase the _trigger_server function attemtps to wrap the socket
using the same SSL Context this fails when we explicitly tell the
context that it will be used on the server side by setting
ssl.Purpose.AUTH_CLIENT

This commit simply removes the wrap_socket and allows any
connection. This is bad for a number of reasons, but the best we
can do here.